requirements.txt
# Direct dependencies used by the current codebase
openai==2.6.1
python-dotenv==1.2.1
pandas==2.3.3
PyYAML==6.0.3
tqdm==4.67.1
